+++ src/client/clientenvironment.h
@@ -79,7 +79,7 @@ public:
 	Client *getGameDef() { return m_client; }
 	void setScript(ClientScripting *script) { m_script = script; }
 
-	void step(f32 dtime, float uptime, unsigned int max_cycle_ms);
+	void step(f32 dtime, double uptime, unsigned int max_cycle_ms);
 
 	virtual void setLocalPlayer(LocalPlayer *player);
 	LocalPlayer *getLocalPlayer() const { return m_local_player; }
